a basic expense tracker

+++ expenses.py
@@ -1,14 +1,31 @@
+import os
 products = []
+
+if os.path.isfile("expenses.csv"):
+	# 讀取舊檔案
+	with open("expenses.csv", "r") as f:
+		for line in f:
+			if "My Expenses Tracker" or "Products" and "Price" in line:
+				continue
+			name, price = line.strip().split(',')
+			products.append([name, price])
+else:
+	print("Create a new file")
+	# 初始建立檔案
+	with open("expenses.csv", "w") as f:
+		f.write("My Expenses Tracker" + '\n')
+		f.write("Products" + "," + "Price" + "\n")
+
+
+# 使用者新增項目
+print("Type 'q' to exit")
 while True:
-	name = input('請輸入商品名稱: ')
-	if name == 'q':
+	name = input("Products: ")
+	if name == "q":
 		break
-	price = input('請輸入商品價格: ')
+	price = input("Price: ")
 	products.append([name, price])
-for p in products:
-	print(p[0], '的價格是', p[1])
-print('已儲存檔案')
-with open('expenses.csv', 'w') as f:
-	f.write('商品' + ',' + '價格' + '\n')
+print("Save file!")
+with open("expenses.csv", "a") as f:
 	for p in products:
 		f.write(p[0] + ',' + p[1] + '\n')
